adding gguf serialization for paratmeters

+++ code/test/MLPTrainingTests.cs
@@ -9,11 +9,11 @@ public class MLPTrainingTests
     public void Training()
     {
         // 4 samples, 3 inputs, binary target (2 classes)
-        var xs = new double[,]
+        var xs = new float[,]
         {
             { 2,  3, -1 },
-            { 3, -1,  0.5 },
-            { 0.5, 1, 1 },
+            { 3, -1,  0.5f },
+            { 0.5f, 1, 1 },
             { 1,  1, -1 }
         };
         var ys = new[] { 0, 1, 0, 1 };  // class indices
@@ -21,7 +21,7 @@ public class MLPTrainingTests
         var brain = new Brain("test", 3, [4, 4, 2], ActivationType.Tanh);
         var input = Operand.Of(xs);  // [4, 3]
 
-        double firstLoss = 0, lastLoss = 0;
+        float firstLoss = 0, lastLoss = 0;
 
         for (var step = 0; step < 50; step++)
         {
@@ -35,10 +35,10 @@ public class MLPTrainingTests
             loss.Backpropagation();
 
             foreach (var m in brain.ParameterMatrices)
-                m.ApplyGradients(0.5, 1.0);
+                m.ApplyGradients(0.5f, 1.0f);
         }
 
         Console.WriteLine($"first_loss={firstLoss:F4}  last_loss={lastLoss:F4}");
-        lastLoss.ShouldBeLessThan(firstLoss * 0.9);  // loss must drop by at least 10%
+        lastLoss.ShouldBeLessThan(firstLoss * 0.9f);  // loss must drop by at least 10%
     }
 }
